+++ asset/verifier.go
@@ -5,6 +5,7 @@ import (
 	"encoding/hex"
 	"fmt"
 	"io"
+	"strings"
 )
 
 var (
@@ -32,8 +33,8 @@ func (v *Sha512Verifier) Verify(rs io.ReadSeeker, desiredSHA string) error {
 		return err
 	}
 
-	if foundSHA := hex.EncodeToString(h.Sum(nil)); foundSHA != desiredSHA {
-		return fmt.Errorf("sha512 '%s' does not match specified sha512: '%s'", desiredSHA, foundSHA)
+	if foundSHA := hex.EncodeToString(h.Sum(nil)); !strings.EqualFold(foundSHA, desiredSHA) {
+		return fmt.Errorf("sha512 of downloaded asset (%s) does not match specified sha512 in asset definition (%s)", foundSHA, desiredSHA)
 	}
 
 	return nil