+++ util/logging/logging.go
@@ -14,9 +14,9 @@ func EventFields(event *types.Event, debug bool) map[string]interface{} {
 	}
 
 	fields := logrus.Fields{
+		"event_id":         event.GetUUID().String(),
 		"entity_name":      event.Entity.Name,
 		"entity_namespace": event.Entity.Namespace,
-		"uuid":             event.GetUUID().String(),
 	}
 
 	if event.HasCheck() {