+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Versioning](http://semver.org/spec/v2.0.0.html).
 - Fixed a bug where the entity API could panic.
 - The agent now respects the HTTP proxy environment variables when using mTLS
 authentication.
+- Improve error message around asset checksum mismatch.
 
 ### [6.1.2, 6.1.3] - 2020-10-28