+++ CHANGELOG.md
@@ -38,6 +38,7 @@ authentication.
 ### [6.1.2] - 2020-10-28
 ### Fixed
 - Fixed a bug where the entity API could panic.
+- Improve error message around asset checksum mismatch.
 
 ### [6.1.2, 6.1.3] - 2020-10-28